plugins: cleanup plugins - replace @Subscribe with @Override - replace events parameters with it - remove redundant @Provides blocks
